Build info (#183)
Include build-info.json in build output. Later this info should be
included in the footer of every page but for now just include the raw
data so we can find out what version is deployed.

The build number logic is changed compared to all other repositories so
that it includes the run attempt and unique build id as suffixes.