Convert mermaid diagrams to svg
The mermaid plugin broke when switching to Yarn. Instead of fixing the
mermaid plugin, the diagrams have been moved to Miro and manually
imported as SVG.

Using this approach has the advantage of not requiring mermaid rendering
in the browser and allows us to review all diagrams in one place.

All of the diagrams have also been reworked with the following
improvements:

* Simplify all sequence diagrams to have minumum amount of actors and
  steps.
* Improve appearance of all diagrams by tweaking the themes and
  introducing line breaks in labels.
* Improve wording of some labels.
* Use better Mermaid variable names.

Additionally, the CSS has been tweaked so the diagrams are legible in dark mode.

Mermaid diagrams are hosted in Miro at https://miro.com/app/board/uXjVNxzJMjE=/.