Add social preview image to html head (#223)
The social preview requires an absolute url for the og:image meta
attribute. The easiest way to achieve this is to hardcode the site base
url to https://docs.immersve.com. As a result, the og:image url is
pointing at an image that does not exist in production yet.

To test before deploy: view source and replace the domain in the
og:image url with the domain for the environment and make sure the image
exists.

To test after deploy: make sure the social preview is working at
https://socialsharepreview.com/?url=https://docs.immersve.com/guides/universal-evm-smart-contract/