Prevent out-of-order heading selection
A hidden ::before element was added on each header to compensate for the
fixed site heading which was hiding content. This extra element overlaps
with the previous paragraph and, since it has a higher natural z-index, makes
the text selection behavior act strangely (the heading would be selected
when selecting text between lines).