Fix display of tables with unwrapable content (#436)
Using display:block for tables makes them ugly when the content is not
long such as with the environments table on the integrator guide.
Restore default display:table for table elements and instead surround
tables with an overflow-x-auto div.

The sandbox account table content is explicitly wrapped with <wbr> tags.